requirements.txt
numpy>=1.24
Pillow>=10.0
PyYAML>=6.0
matplotlib>=3.7
scipy>=1.10
torch>=2.1
tqdm>=4.66
pytest>=7.4
